Reworked configuration hierarchy
