+++ packages/library-base/src/commonMain/kotlin/io/realm/BaseRealm.kt
@@ -21,13 +21,13 @@ interface BaseRealm : Versioned {
     /**
      * Configuration used to configure this Realm instance.
      */
-    val configuration: RealmConfiguration
+    val configuration: Configuration
 
     /**
      * Returns the current number of active versions in the Realm file. A large number of active versions can have
      * a negative impact on the Realm file size on disk.
      *
-     * @see [RealmConfiguration.Builder.maxNumberOfActiveVersions]
+     * @see [Configuration.Builder.maxNumberOfActiveVersions]
      */
     fun getNumberOfActiveVersions(): Long
