+++ packages/library-base/src/commonMain/kotlin/io/realm/RealmObject.kt
@@ -94,7 +94,7 @@ public fun RealmObject.isValid(): Boolean {
  * object. If the observed object is deleted from the Realm, the flow will complete, otherwise it will
  * continue running until canceled.
  *
- * The change calculations will on on the thread represented by [RealmConfiguration.notificationDispatcher].
+ * The change calculations will on on the thread represented by [Configuration.notificationDispatcher].
  *
  * @return a flow representing changes to the object.
  */
