+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/BaseRealmImpl.kt
@@ -27,7 +27,7 @@ import kotlin.reflect.KClass
 
 @Suppress("UnnecessaryAbstractClass")
 abstract class BaseRealmImpl internal constructor(
-    final override val configuration: InternalRealmConfiguration,
+    final override val configuration: InternalConfiguration,
     dbPointer: NativePointer
 ) : BaseRealm, RealmStateHolder {
