+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/InternalConfiguration.kt
@@ -16,7 +16,7 @@
 
 package io.realm.internal
 
-import io.realm.RealmConfiguration
+import io.realm.Configuration
 import io.realm.RealmObject
 import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.SchemaMode
@@ -25,10 +25,10 @@ import kotlin.reflect.KClass
 
 /**
  * An **internal Realm configuration** that holds internal properties from a
- * [io.realm.RealmConfiguration]. This is needed to make "agnostic" configurations from a base-sync
+ * [io.realm.Configuration]. This is needed to make "agnostic" configurations from a base-sync
  * point of view.
  */
-interface InternalRealmConfiguration : RealmConfiguration {
+interface InternalConfiguration : Configuration {
     val mapOfKClassWithCompanion: Map<KClass<out RealmObject>, RealmObjectCompanion>
     val mediator: Mediator
     val nativeConfig: NativePointer
