+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/RealmSchemaLoweringExtension.kt
@@ -16,9 +16,9 @@
 
 package io.realm.compiler
 
+import io.realm.compiler.FqNames.REALM_CONFIGURATION
 import io.realm.compiler.FqNames.REALM_CONFIGURATION_BUILDER
-import io.realm.compiler.FqNames.REALM_LOCAL_CONFIGURATION
-import io.realm.compiler.FqNames.REALM_SYNC_CONFIGURATION
+import io.realm.compiler.FqNames.SYNC_CONFIGURATION
 import io.realm.compiler.FqNames.SYNC_CONFIGURATION_BUILDER
 import io.realm.compiler.Names.REALM_CONFIGURATION_BUILDER_BUILD
 import io.realm.compiler.Names.REALM_CONFIGURATION_WITH
@@ -84,7 +84,7 @@ class RealmSchemaLoweringExtension : IrGenerationExtension {
                 @Suppress("LongMethod")
                 override fun visitCall(expression: IrCall): IrExpression {
                     val name = expression.symbol.owner.name
-                    if ((expression.type.classFqName == REALM_LOCAL_CONFIGURATION || expression.type.classFqName == REALM_SYNC_CONFIGURATION) && name in setOf(
+                    if ((expression.type.classFqName == REALM_CONFIGURATION || expression.type.classFqName == SYNC_CONFIGURATION) && name in setOf(
                             REALM_CONFIGURATION_BUILDER_BUILD,
                             REALM_CONFIGURATION_WITH
                         )
@@ -110,7 +110,7 @@ class RealmSchemaLoweringExtension : IrGenerationExtension {
                                 }
                         }
                         val buildFunction =
-                            if (expression.type.classFqName == REALM_SYNC_CONFIGURATION) {
+                            if (expression.type.classFqName == SYNC_CONFIGURATION) {
                                 // We only reach this if we are building a sync configuration so calling "!!" is safe
                                 syncBuildFunction!!
                             } else {
