+++ test/base/src/androidTest/kotlin/io/realm/test/shared/SchemaTests.kt
@@ -20,6 +20,7 @@ import io.realm.RealmObject
 import io.realm.entities.Sample
 import io.realm.entities.link.Child
 import io.realm.entities.link.Parent
+import io.realm.internal.InternalConfiguration
 import io.realm.internal.interop.Table
 import kotlin.reflect.KClass
 import kotlin.test.Test
@@ -96,7 +97,6 @@ class SchemaTests {
 
     private val RealmConfiguration.companionMap: Map<KClass<out RealmObject>, io.realm.internal.RealmObjectCompanion>
         get() {
-            @Suppress("invisible_member")
-            return (this as io.realm.internal.RealmConfigurationImpl).mapOfKClassWithCompanion
+            return (this as InternalConfiguration).mapOfKClassWithCompanion
         }
 }
